Completed the basic functions

+++ controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require('../models/contactModel');
 
 /* Get all contacts
 @route GET /api/contacts
-@access public */
+@access private */
 
 const getAllContacts = asyncHandler(async (req, res) => {
-  const allContacts = await Contact.find();
+  const allContacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(allContacts);
 });
 
 /* Get contact by ID
 @route GET /api/contacts/:id
-@access public */
+@access private */
 
 const getContactById = asyncHandler(async (req, res) => {
   const contactId = req.params.id;
@@ -27,7 +27,7 @@ const getContactById = asyncHandler(async (req, res) => {
 
 /* create new contact
 @route POST /api/contacts
-@access public */
+@access private */
 
 const createContact = asyncHandler(async (req, res) => {
   const { name, phone, email } = req.body;
@@ -36,7 +36,12 @@ const createContact = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error('All fields are required! ');
   }
-  const newContact = await Contact.create({ name, email, phone });
+  const newContact = await Contact.create({
+    user_id: req.user.id,
+    name,
+    email,
+    phone,
+  });
   res.status(201).json({
     message: 'Contact added successfully!',
     new_contact: newContact,
@@ -45,7 +50,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 /* Update contact
 @route PUT /api/contacts/:id
-@access public */
+@access private */
 
 const updateContact = asyncHandler(async (req, res) => {
   const contactId = req.params.id;
@@ -57,12 +62,20 @@ const updateContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error('Contact not found');
   }
+  //Checking if user has permission to update
+  if (contactFound.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      "User don't have permission to update other user's contact"
+    );
+  }
 
   //updating if it exists
 
   const updatedContact = await Contact.findByIdAndUpdate(contactId, req.body, {
     new: true,
   });
+
   res.status(200).json({
     message: 'Contact updated successfully!',
     updated_contact: updatedContact,
@@ -71,7 +84,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 /* Delete contact
 @route DELETE /api/contacts/:id
-@access public */
+@access private */
 
 const deleteContact = asyncHandler(async (req, res) => {
   const userContact = await Contact.findById(req.params.id);
@@ -80,6 +93,14 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
 
     throw new Error('Contact not found!');
+  }
+
+  //Checking if user have permission to delete
+  if (userContact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      "User don't have permission to delete other user's contact"
+    );
   }
 
   await Contact.findByIdAndDelete(req.params.id);
